+++ services/account/account_server-batch/src/main/resources/application-local.yml
@@ -1,9 +1,9 @@
 spring:
   datasource:
     driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://${local-db.host}:${local-db.port}/${local-db.name}
-    username: ${local-db.username}
-    password: ${local-db.password}
+    url: jdbc:postgresql://${local.domain-db.host}:${local.domain-db.port}/${local.domain-db.name}
+    username: ${local.domain-db.username}
+    password: ${local.domain-db.password}
 
   jpa:
     properties:
@@ -14,12 +14,6 @@ spring:
         default_batch_fetch_size: 100
     open-in-view: false
     show-sql: true
-
-  batch:
-    job:
-      enabled: false
-    jdbc:
-      initialize-schema: always
 
 logging:
   level:
